+++ SharePointFramework/@Shared/package.json
@@ -1,6 +1,6 @@
 {
   "name": "pp365-shared",
-  "version": "1.7.2-868.1",
+  "version": "1.7.2-868.3",
   "repository": {
     "type": "git",
     "url": "https://github.com/Puzzlepart/prosjektportalen365.git",
