+++ SharePointFramework/@Shared/src/services/PortalDataService/GetStatusReportsOptions.ts
@@ -1,0 +1,7 @@
+export type GetStatusReportsOptions = {
+  filter?: string
+  top?: number
+  select?: string[]
+  publishedString?: string
+  useCaching?: boolean
+}
