+++ SharePointFramework/@Shared/src/services/PortalDataService/index.ts
@@ -20,18 +20,12 @@ import {
   SPProjectColumnItem,
   StatusReport
 } from '../../models'
+import { GetStatusReportsOptions } from './GetStatusReportsOptions'
 import {
   IPortalDataServiceConfiguration,
   PortalDataServiceDefaultConfiguration,
   PortalDataServiceList
 } from './IPortalDataServiceConfiguration'
-
-export type GetStatusReportsOptions = {
-  filter?: string
-  top?: number
-  select?: string[]
-  publishedString?: string
-}
 
 export class PortalDataService {
   private _configuration: IPortalDataServiceConfiguration
@@ -110,13 +104,14 @@ export class PortalDataService {
   }
 
   /**
-   * Get project status sections
+   * Get project status sections using caching.
    */
   public async getProjectStatusSections(): Promise<SectionModel[]> {
     try {
       const items = await this.web.lists
         .getByTitle(this._configuration.listNames.STATUS_SECTIONS)
-        .items.get()
+        .items.usingCaching()
+        .get()
       return items.map((item) => new SectionModel(item))
     } catch (error) {
       return []
@@ -405,7 +400,8 @@ export class PortalDataService {
   }
 
   /**
-   * Get status reports
+   * Get status reports using caching by default. Can be turned off by setting
+   * `useCaching` to `false`.
    *
    * @param options Options
    */
@@ -413,7 +409,8 @@ export class PortalDataService {
     filter = '',
     top,
     select,
-    publishedString
+    publishedString,
+    useCaching = true
   }: GetStatusReportsOptions): Promise<StatusReport[]> {
     if (!this._configuration.siteId) throw 'Property {siteId} missing in configuration'
     if (stringIsNullOrEmpty(filter)) filter = `GtSiteId eq '${this._configuration.siteId}'`
@@ -425,6 +422,7 @@ export class PortalDataService {
         .orderBy('Id', false)
       if (top) items = items.top(top)
       if (select) items = items.select(...select)
+      if (useCaching) items = items.usingCaching()
       return (await items.get()).map((i) => new StatusReport(i, publishedString))
     } catch (error) {
       throw error
