+++ SharePointFramework/@Shared/src/util/conditionalClassName.ts
@@ -4,5 +4,5 @@
  * @param classNames Array of class names
  */
 export function conditionalClassName(classNames: string[] = []): string {
-    return classNames.filter(Boolean).join(' ')
+  return classNames.filter(Boolean).join(' ')
 }
