+++ SharePointFramework/PortfolioWebParts/package.json
@@ -45,7 +45,7 @@
     "moment": "2.24.0",
     "msgraph-helper": "0.7.3",
     "object-assign": "4.1.1",
-    "pp365-shared": "1.7.2-868.1",
+    "pp365-shared": "1.7.2-868.3",
     "react": "16.13.1",
     "react-calendar-timeline": "0.27.0",
     "react-dom": "16.13.1",
