+++ SharePointFramework/ProjectWebParts/package.json
@@ -45,7 +45,7 @@
     "get-value": "3.0.1",
     "interactjs": "1.6.2",
     "msgraph-helper": "0.7.1",
-    "pp365-shared": "1.7.2-868.1",
+    "pp365-shared": "1.7.2-868.3",
     "pzl-spfx-components": "0.0.11",
     "react": "16.13.1",
     "react-calendar-timeline": "0.26.0",
